requirements.txt
streamlit
requests
beautifulsoup4
spacy
python-dotenv
google-generativeai
python-docx
PyPDF2
weasyprint
fpdf
langdetect
easyocr
pymupdf
streamlit-webrtc
av
pyarrow
soundfile
streamlit-extras
flask_cors
razorpay

# python -m spacy download en_core_web_sm
# for pymupdf to work properly, you may need to install additional system dependencies:
# sudo apt update
# sudo apt install libpango-1.0-0 libpango1.0-dev
# streamlit run login.py


# python -m spacy download en_core_web_sm
# for pymupdf to work properly, you may need to install additional system dependencies:
# sudo apt update
# sudo apt install libpango-1.0-0 libpango1.0-dev

